+++ unit-tests/ValidationTest.php
@@ -27,7 +27,8 @@ use Phalcon\Validation\Validator\PresenceOf,
 	Phalcon\Validation\Validator\StringLength,
 	Phalcon\Validation\Validator\Email,
 	Phalcon\Validation\Validator\Between,
-	Phalcon\Validation\Validator\Url;
+	Phalcon\Validation\Validator\Url,
+	Phalcon\Validation\Validator\CreditCard;
 
 class ValidationTest extends PHPUnit_Framework_TestCase
 {
@@ -205,6 +206,71 @@ class ValidationTest extends PHPUnit_Framework_TestCase
 		$messages = $validation->validate($_POST);
 
 		$this->assertEquals(count($messages), 0);
+	}
+
+	public function providerCreditCardNumberValid()
+	{
+		return array(
+			array('378282246310005'), //amex
+			array('4012888888881881'), //visa
+			array('38520000023237'), //dinners
+			array('5105105105105100'), //mastercard
+			array('6011000990139424') //discover
+		);
+	}
+
+	/**
+	 * @dataProvider providerCreditCardNumberValid
+	 */
+	public function testValidationCreditCardValid($number)
+	{
+		$_POST = array('number' => $number);
+
+		$validation = new Phalcon\Validation();
+
+		$validation->add('number', new CreditCard());
+
+		$messages = $validation->validate($_POST);
+
+		$this->assertEquals(count($messages), 0);
+	}
+
+	public function providerCreditCardNumberInvalid()
+	{
+		return array(
+			array('1203191201121221'),
+			array('102030102320'),
+			array('120120s201023'),
+			array('20323200003230'),
+			array('12010012')
+		);
+	}
+
+	/**
+	 * @dataProvider providerCreditCardNumberInvalid
+	 */
+	public function testValidationCreditCardInvalid($number)
+	{
+		$_POST = array('number' => $number);
+
+		$validation = new Phalcon\Validation();
+
+		$validation->add('number', new CreditCard());
+
+		$messages = $validation->validate($_POST);
+
+		$expectedMessages = Phalcon\Validation\Message\Group::__set_state(array(
+			'_messages' => array(
+				0 => Phalcon\Validation\Message::__set_state(array(
+					'_type' => 'CreditCard',
+					'_message' => 'Field number is not valid for a credit card number',
+					'_field' => 'number',
+					'_code' => '0',
+				))
+			)
+		));
+
+		$this->assertEquals($expectedMessages, $messages);
 	}
 
 	public function testValidationIdenticalCustomMessage()