+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Now Phalcon\Mvc\Model::__call invokes finders as in __callStatic
 - Fixed `Phalcon\Db\Dialect\Postgresql::getColumnDefinition` for `BIGINT` and `BOOLEAN` data types
 - Fixed `BOOLEAN` default value in `Phalcon\Db\Dialect\Postgresql`
+- Added `Phalcon\Validation\Validator\CreditCard` - validation credit card number using luhn algorithm
 
 # [2.0.7](https://github.com/phalcon/cphalcon/releases/tag/phalcon-v2.0.7) (2015-08-17)
 - `Image\Adapter\Gd::save()` no longer fails if the method or the instance is created with a filename without an extension